chore: Fix dependency updater script

+++ update.sh
@@ -1,14 +1,35 @@
 #!/bin/bash
 
+# make sure if its in master branch
+if [[ $(git rev-parse --abbrev-ref HEAD) != "master" ]]; then 
+    echo 'Command must be run in master branch';
+    exit 1
+fi 
+# make sure the master branch is clean 
+if ! git diff-index --quiet HEAD --; then
+    # No git changes (no update)
+    echo 'Master branch is dirty, revert or commit your changes first'
+    exit 1
+fi
 # update root package
 ncu -u
 # update pckages
 for file in packages/*/package.json; do
     ncu -u --packageFile "$file"
 done
+# check if update successful
+if git diff-index --quiet HEAD --; then
+    # No git changes (no update)
+    echo 'Packages already up to date!'
+    exit 1
+fi
 # remove all dependent libraries
 rm -f yarn.lock 
 rm -rf node_modules
 # install and test
 yarn install
-yarn test
+# push to upgrade branch
+git branch -D upgrade 
+git checkout -b upgrade 
+git add . && git commit -m "chore: Upgrade dependencies"
+git push -u origin upgrade
